Adjusted the cubeCreator's vertices and UVs
When creating the vertices and UVs of the cube, the 3rd material will use a sepperate list of vertices adn UVs so that we can feed the shader the proper UV values only to the thrid material. The idea will be to set the UVs of the third material to range from 0 to 1, where 0 is the closest to the center and 1 is furthest (ie a corner).