The cube's edges now use a shader to fade from one texture to the other
The cube's edges use a custom unlit surface shader that changes the texture depending on the UV value of the given point. This requires that the box sets the UV values so that the value 1 is the vertices which make up the center square while 0 is the vertices on the cube's corners. This produces the desired fade effect where we can see a black to white fade as the texture goes across the edge.