Fixed the dynamic batching error
When setting the triangles and materials of a cube, it was looking at the length of the same triangle array 3 times, instead of looking at each of the 3 triangle arrays sepperatly.

A new issue has been found: The edge mesh of a cube has wierd rendering issues when there is another cube bellow it. Sometimes, the edge renders behind an object.