Can now flip the edge's blend texture
The edges can now flip the texture that is being blended with a public boolean value. This this allows the cube to blend from dark in the center to bright on the edges, but this will not switch the main and second material automatically.

Also, a wierd rendering issue occurs with the cube's edges in the main puzzle map. The lighting on it is wierdly bright, and this does not occur in the test area.