Fixed the cube edge lighting issue. Updated the edge flip blend method.
The cube's edge shader was using ambient lighting from the environment, so the parameter noambient was added to #pragma. Also, the way the shader determines when to flip the blend texture is done by the second set of UV's Y value. It still uses the X value as the main fade value.