Addded edges to many cubes in the puzzleRooms
implemented the edge fade for the cubes is a few puzzleRooms. The final room has all it's cubes use the edge fade effect and looks much better now. PuzzleRoom 3 has it's flooring use an edge to signify that the cube is touching the wall, but the effect is also used to represent a drop off to entice the player to approach it. Maybe, to make it more obvious that a floor is reaching the wall, have another shader effect come into play if the surface is close to the room's edges?