Re-formatted puzzleRoom5
PuzzleRoom 5 now seems a little less complex. The desired puzzle solution will still be the same, but for now the portals are not in their proper rotation. The layout of the room is what is final.

The memory leak issue has not been found and will required further testing with a complete build.