Reposition the puzzleRooms past 5 to reflect it's new final size. Adjusted edges on rooms 1-4
with puzzleRoom 5 being final, we can now fix the size of it's next waitingRoom since it's sizes won't be changing anymore. Also, the edges of the cubes in the previous rooms have been updated to reflect a new consitency (Paths that allow or recommend the player to walk off their side will have edges that fade into the darkened texture).