PuzzleRoom 6 has it's base layout adjusted
The new idea behind puzzleRoom 6 will make it a bit more obvious the player needs to land on a specific position. Also, once the player lands on the right position and falls into the hole to properly set their rotation, they will walk through the hall that leads to the exit, letting them enter the next waitingRoom at the right rotation.