PuzzleRoom 6's top half has all it's walls and edges properly set
The top side of the puzzleRoom 6 now uses the fade on all it's cubes. The Idea conveyed by this room is that the player does not need to walk onto the darkened floor. The bright marble flooring should hopefully guide the player. Once they are falling into the bottom half of the room, they should be able to see that they are required to land on the bright flooring and not the dark.