Adjusted the PuzzleRoom 6's side swap portal
The portal used to go between the top and bottom half of the room has been adjusted a bit. Now, the cubes around it are not waffer thin and the player can enter the portal when upside down on the bottom half to quickly return to the top half if needed. Ths point of this portal is for the player to be able to go backwards into room 4 when they are making their way back to the startingRoom with the correct rotation.