Fixed the PuzzleRoom 6's bottom half path
The bottom half of the room now has it's path set. The player will fall through the hole and be required to go through a few jumps to reach the exit on the proper rotation. This room might be too dificult, but it can be simplified by moving the portal to connect between the pillars and preventing the player from leaving the bottom half of the room once they reach there.