PuzzleRoom 7 is now getting edge treatement
The puzzleRoom 7 is now having cube edges implemented. So far, the first main portal and the main floor have had edges implemented. The floor is now comprised of 10 cubes so that the edge effect can properly transition from portals and give a nice visual effect on the portals in the corners of the room.