Fixed the rotation of the cubes in puzzleRoom 7
Most of the cubes that formed the puzzleRoom 7 had a rotation on themselves to fit their position. This update reverts all their rotations back to a zero angle but updates their sizes to look like nothing has changed. This allows us to easily modify the edges of the cubes as every cube has the same face direction (ie, All left faces face left).