PuzzleRoom 7's portal's cubes now use fade effect
The cubes that are used to transition between the portals of puzzleRoom 7 now have fade effects on them. So far only the cubes that are touching the portals have edges as there will be a lot of edges and splitting cubes into pieces if we were to complete the whole room.