Updated puzzleRoom 7's main portal and exit edges
The main portal of the room has it's edges updated to look more unique than all the other portals in the game. The exit path of this room also has it's cube's edges updated. They seem fine but may need a bit of changes in the end ocne the whole room is updated.