Added more edges to the puzzleRoom 7's cubes
The main path, the front wall along with it's hall and a bit of the main floor portals have had edges added to give the player a sense of direction. The edges will show where the player can go by following the white part of the cubes. All edges that lead to a drop or a dead end of such should have darkened edges. If a drop leads to progress/a new path, have it end with a bright texture.