Fixed the above/bellow exit/entrance wall
The wall's second UV set is now properly set for the above/bellow walls on the exit/entrances. The cahnge involves the inclusion of the wal's transform when setting the vertices and UVs. This is to access it's final position in the world, letting us properly place the UV2 values in the room.