Fade now goes both up and down. Also added two lengths and a offset as customizable values.
The fade effect now goes both up and down the wall. Three new values have been added:

fadeAmount will stretch the distance of the actual fading. If it's 3, then it takes 3 unity units to fade from 1 texture to the next.

preFadeLength marks how much distance from the center of the fade the texture needs to travel before it begins the fade effect.

fadeOffset will change where the center of the fade starts. This will however require a new implementation of how the walls and their vertices are placed. This may be scrapped later.