Added textures to the walls. Fixed the wall's UVs
Added textures to the walls to test the UVs. Fixed how the UVs were set so that both entrance and exit seamlessly connect between each wall object. Saturated the blend value in the shader to prevent UV values above 1 or bellow 0. Added a new variable to the puzzleRooms that controls the scaling of the wall texture, WallStretch.