Wall shader now uses a gradient to generate more dynamic visuals
The puzzleRoom walls now require the use of 3 textures:

The expected normal texture that will be the main wall texture.

A second texture to be used that visually fades out as the wall progresses away from the center.

A 3rd repeating noise texture that is used as a gradient to control the visual randomness of the wall. This includes tilling of a second layer of the first texture and how the second texture is applied over the first.