Fixed the UVs of the walls above and bellow exit and entrances by introducing a UVYOffset variable
When handling a wall above or bellow and exit/entrance to a puzzleRoom that is not centered, an offset relative to the exit/entrance's position from the center of the room is brought into the wall's UV4. This is to ensure the wall's still have a mostly consistent UV height when placed next to the left/right exit/entrance walls.

Also ajusted the fade values of each room.