Minor fix to the above exit/entrance wall's UVs. Added more textures
The wall's that are placed above the entrance and exits still do not have an accurate UV placement if the entrance or exits are not on the center, but have now been adjusted to be a bit more accurate. The desired equation has not yet been found.

Added more textures to the game top use on the walls as either textures or noise functions.