Player now does not reset upon startup. Opens the window for an intro.
On startup, the player does not run the ResetPlayer function. A new SetupPlayer function will properly place the player and run any upcomming functions that will be used when the built in Start() functions start running.

This will be useful when setting up an intro, as it lets us to start the player in an immobile or special menu state.