Added a new state and state "type" to the player
The player now has 3 state "types": grounded, airborn, and "in menus". The menu type so far only has one state: Intro.

The Intro state is the state the player stats in on startup. It is used for the intro of the game. Currently, the player cannot move their position or camera in the Menu state type. Other functions may be changed, as their legs are not being fired.