Player now begins in a proper position and state
On startup, the player will begin in the Intro state. With this state being a part of the "in menus" states, the player will run any leg ray traces or handle movement inputs. This will allow the player to remain still as the intro animates itself.

Once the player changes to a standing state, their camera is already at the desired height so that a state change that introduces leg rays will not move the camera from it's current spot (assuming the camera is at it's default spot set by AdjustCameraPosition when in the previous state)