Added animation to the intro state.
When in the intro state, the player's camera will animate itself. The player's camera will follow a straight path from the outside area linked to the starting room to the player's position. It determines the position of the camera by firing a ray from the player's default camera position forwards by a set amount. The ray that is fired will ignore all but teleport colliders, causing it to teleport due to the startingRoom's large window.

So far, the animation will always run at startup. The player will enter the standing state once the camera is finished it's animation, giving them control to move around.

Note: During the camera's animation, it may place the camera close enough to the portal so that the camera clips the portal but not the wall behind it. This will require altering the animation so that the camera is never too close to the portal, which requires knowing how far the player is from the portal in the first place.