Fixed portals to properly layer their triggers. Fixed the camera culling the portal issue.
All portal triggers now set themselves to the "Portal Trigger" layer. After this was applied, the problem where the camera was clipping the portal did not occur anymore.

Also, added a check when handling ray casts: If the distance is reduced to or less than 0 by the anti-infitnite loop protection, print to the console as this may be reason behind the "teleport but reach behind the place that we teleported to" issue.