StartingRoom window is now placed relative to ground height
The exit window of the startingRoom can now be placed relative to the height of the terrain directly bellow it. This allows us to reposition the window and update it's height in real time.

Also editted the terrainController's start function so that the startingRoom will run it's startup before it parses the terrainHeight.