Added another menu state. Player camera can now render terrain when outside in menus
There are now two states used for the "Menus" state type: InMenu and LeavingMenu.

InMenu so far fires a ray from the player past the startingRoom's window and into the outside area. It remainds immobile, until prompted to change to the LeavingMenu state. The LeavingMenu state simply animates the camera returning to the player, giving them control once it's complete.

Also added a PlayerRenderTerrain helper function as the player camera was not rendering the terrain since we disabled it on startup. Now the player will render the terrain if the camera is fired through the startingRoom window.