Added animation to the window while in the "InMenu" state.
While in the InMenu state, the startingRoom's outside window will strafe along the X axis. This will make it seem like the camera is animating across the terrain.

However, when the window's Z axis changes, it's height while parsing the noiseProvider is incorrect. It originally was correct as it's Z position was 0.