Added a basic menu to the game with a working start button
A new script called Menu will be linked to the player. It will control all GUI elements and has access to the CustomPlayerController in order to send commands once it's buttons get clicked.

Currently, pressing the Start button will send a command to the player that the start button has been pressed. This will start moving the player from the InMenu state to the LeavingMenu state, making this a fully working menu.