Added an animation to the button when clicked. Added MouseHover for start button
When the button is pressed, a variable gets ticked and an animation on the button starts to make it fade away. This effect does not look great, but the format of an animation is now in place.

Also, the event of the mouse hovering over the start button is now caught and handled using event triggers for the button.