Added a Quit button along with it's event functions
A new Quit button has been added to the menu. It is placed bellow the start button and has it's event functions currently working. The application does not quit when clicking the button but instead prints to the console. While in the MainToIntro transition state, the button slides off the left side of the screen instead of "exploding" like the start button.