Added a panel to the canvas along with hierarchy adjustments and a helper function.
Panels are now new objects that can be added to the scene. So far, a Cover panel is added and is used when the game is ending. It will cover the screen with a black color during the transition and will be fully black before the game exits.

The canvas uses the AdjustRatio helper function which will allow us to adjust when a value with a range of [0, 1] is re-intepretted given a new min and max range (also within the [0, 1] range). It will be useful at deciding when an object should start animating. So far, the cover panel will only start to fade in 25% through the quitting transition and will fully block the screen 80% through.