Start button now transforms to the Continue button once the game has started
On startup, the game will start in the menu's Startup state. Pressing the START button will bring the menu and the player to the "intro" state (for the menu, it just animates to the empty state). After the menu is done it's animation, the start button will update itself to become the "continue" button, which will now simply close the menu instead of bringing the player back to the intro.

Also added a new helper function used to extract a [0, 1] range from a set of timeRemaining and timeMax values.