Added the Escape key to enter and exit the menu.
The player can exit and enter the menu by pressing the escape key. This is handled by the PlayerController. It currently does not put the player into the menu state, but it does update the menu's states. Also, a new state of MainToEmpty has been added. It is used to animate the menu closing.