Added a whole new transition state implementation
Transition states will now have a Transition object that is used to track their current transition time and procedding state. If a transition ends up not transitionning into another state or requires a more unique approach, such as quitting the game, it will simply transition into it's current state and the condition will be handled in the ChangeState function in the (state == newState) scenario.

Also, a new StateFunction object is used to track each UI element and what function they will run for each state. This is to simplify the update function and make it easier to introduce new functions as you only need to write the required update function then add it to the proper StateFunction array (and update the StateFunctionInit function).