Fixed the player states when leaving the intro and opening the menu
When exiting the menu for the first time and putting the player into the intro sequence, the player will be using their two intro states. This is the only time they will ever use them as any uses of the menu later will be done with varaibles as they limit certain actions. We will always keep the player in their current state when the menu is open. So far, when pressing the escape key, the menu will open but the player will still have control.

Also updated a few values pertinent about the visuals of the menu during the intro