Added new Sensitivity states accesed using the Sensitivity button
Pressing the sensitivity button will transition the menu into the Sensitivity state, which does nothing for now. It will keep the sensitivity button on-screen in this state, letting the user press it to go back to the main state. Pressing escape will also leave the sensitivity menu. The planned use of this menu

Also, the Sensitivity button will always seem to be hovered over while in the sensitivity menu.