Added elements to the sensitivity panel
The sensitivity panel now adds UI elements to itself on setup. This includes text and a slider which can be interacted with and runs a function on it's value update. This will be improved upon to actually effect stats from the player.

Note the slider is cloned from a reference to another slider object that is already linked to the menu script.