Added a rotation animation to the buttons for the video states
When entering and leaving the video state, the other menu buttons will undergo a rotation as they move offscreen. This rotation changes the button's pivot point, which will effect how the button is placed, so when the VideoToMain update function runs it's final call, it will reset the pivot point back to the regular (0.5, 0.5) so the button can be properly placed.