Added smaller panels in the Video panel
The Video panel now has smaller panels within. These panels are split into text objects on their left and interractable UI elements on their right. So far, a dropdown menu is added to the first panel. These panels and interractable elements will be used to control the video/graphics of the game.