Added UI elements and functions to each option panel
Each option panel in the video panel now has an element on the right side that calls a function when interacted with. Their visual layout is controlled by the reference they are cloning, so adjusting the object being references in the editor will effect the component in the menu when the game starts.