Windowed mode, mouse lock and pause on focus loss are all working
The options in the video menu that control the windowed mode of the screen, the state of the mouse and whether it should be locked into the screen (It should however always be locked while out of the menu) and how the game reacts once it loses focus are now working properly with a built game. The editor cannot control the  window mode or mouse lock state so testing these requires building the game.