Implemented resolution change and framerate change
The options to change the resolution and framerate are now working. However, when changing the framerate, the testing of the mouse's sensitivity requires a change as it updates using Update yet doesnt use Time.deltaTime, meaning different farmerates will have different speeds. So, it has been changed to be the same for each framerate