The quit button now has added animation
When pressing the quit button, it will start undeergoing effects as the quit value approaches it's max, which will cause the game to enter the quitting state. The animations the quit button undergoes (color, size and position change) will also return the button to it's original state if the user does not make the quitting value reach it's transition limit.