Player can now skip the intro using the escape key. The sun is now placed as an in-world object
Pressing the menu key (escape) while the player is going through the intro, they can now skip the intro.

The terrainController now also controls the lights if the scene, which have been sepperated into two objects: Direction and point. The point will be used as the physical object of the sun, which includes it's flare. The directionnal light is the actual lighting of the scene.