StartingRom's window now renders sun flare
The sunflare can now be seen from the startingRoom's window. This required a large series of functions that run down from the startingRoom down to the PortalView, letting us call a AddFlareLayer to add the flayerlayer to all cameras linked to a portal set.

The sun's position has also been updated to use 3 variables to determine where it will be placed.

Also, fixed a bug with the main menu where pressing escape gives the player control and leaves the menu in main