Added flare intensity, which the menu will modify
The sun flare's intensity is now modifiable, which the menu will now use to it's advantage. The new startup will mess with the sun's intensity. Currenty, the intensity of the flare is reduced back to normal as the menu's startup ends.

The sun's placement will now require a bit of modification: Since the flare will not show properly if it's too far away, the sun will now be placed a given distance away from the focus point. This will, however, require the focus point to be changed to be the camera instead of the window or the player object (As movements of the camera and not the player, such as falling or walking around the window, will show how the sun is very close).