Started implemention the changes to the focus point. Fixed a couple bugs.
Started implemention changes to the terrainController's focus point. The focus point should not be the player's base nor the window's position but instead the player's camera and the window's portal's camera used to render it's portal. This is to ensure the sun can be very close to the player without seeming to move.

Fixed a bug that sometimes had the sensitivity change when clicking into the game screen (what is causing this? Is the user clicking the mouse onto the slider UI element in the world?) This problem has caused us to change the video and sens panels to be inactive until needed.

Fixed the bug that prevented skipping the intro when in the LeavingIntro state (did not take into account when the user presses escape later in the intro during the LeavingIntro state)