OutsideTerrain's focus point is now updated on the fly using new references
The outside terrain will now determine which focus point will be used by examining where the player's camera is positionned. It takes into account the outside window's position (from a reference set by the startingRoom) and compares it to the player camera. This does seem like a hack as if the player does end up on a Z position that is lower/less than the window's Z position, it will focus on the player's camera. But, this occurence should not occur as the room's positions have it so the puzzles work towards the positive Z axis while the outside works towards the negative.