Fixed a few bugs involving the starting sequence of the game
Adjusted how the player updates their camera's near and far clipping planes when animating the intro sequence and when the camera stops using the teleporter from the startingRoom's window.

The bug involving the startingRoom's window's camera not properly rendering the outside terrain is very strage and I cannot find a way to fix it yet. After the intro (and during), the window's portal does not render the portal's view with a proper render distance. This issue is only active until the camera updates it's view in CERTAIN cases. Currently, after the intro is complete, if the user is to more the camera or themselves using the WASD keys, the camera will update and fix the render distance. HOWEVER, jumping WILL NOT update the camera's render distance. Also, unlocking the player camera from the playerController and moving the camera around WILL NOT update the render distance.

The final thing to note is that every portal camera does infact have a very large render distance, and printing the camera's stats before it renders using the PortalView script proves this. However, the camera will still have the wrong render distance.

So far I believe the issue does not have something to do with the portalView or any portal script specifically, but involves the PlayerController or some function calling to a portal script.