Fixed a few bugs involving the intro sequence and the player camera's rendering
The problem with the startingRoom's portal not properly rendering the terrain was due to the fact that the startingRoom's portal's cameras use the playerCamera's near and far clipping planes. This is actually wierd because testing it with other portals indicates the other portals act the same. This issue is solved by not reducing the player's rendering distance when they change their rendering layers.

Also, when the player camera is returning to it's default position from from intro, prevent the camera from occupying a spot that is very close to the window as having it placed close enough to the window's portal will cause rendering issues for a single frame.