Added a loading bar to the main menu
The terrainController can now track how far into the terrain loading it currently is. This allows us to display and implement a loading bar. The startup of the intro will feature a loading bar that shows the current progress of the loading. While the terrain is not fully loaded, the startup transition's remainingTime will continuously get reset.

Note: During the startup, we may want to keep the player from moving the window as it may move the window into a new chunk, causing some chunks to unload and  bring the loading bar back a bit.