Adjusted terrain startup. Working on improving the Dropdown menus
The terrain will now not be forced to generate all the terrain before startup. With the inclusion of the loading bar, we can spend more time waiting for the terrain to generate in the startup menu state rather than the unity initialization state before the game even runs it's first Update call.

The dropdown menus are now using a new background and text format. Currently the dropdown menu's visual aspect is fine but it's sizes are not properly updated.