The sizes of the dropdown boxes have been adjusted to match the screen size
The dropdown boxes and their contents will now resize relative to the screen size, along with the text's outlines. The starting sizes of the original reference dropdown box has also been adjusted to remove the excess borders that it applied due to it's original borders.

Note: A lot of the dropdown box size changes use the child order of the hierarchy to get the proper components, so removing or adding a new object to the dropdown reference will cause issues involving the dropdown box's size.