Improved the transitionning between inside and outside.
When the player falls out of the starting room and into the outside, there needs to be a seamless transition. In the current state, the act of falling out of the window is now seamless. This introduces an UpdateRenderingLayer() function in the playerController which currently runs twice a frame. A problem however is that looking out the starting window will show the sunflare is blocked. This is due to a collider that should be set to active and inactive from running UpdateRenderingLayer(). The goal currently is to find out when  to run UpdateRenderingLayer() so that we can get a seamless transition AND let the sun flare be viewable.

Fixed the playerCamera not properly detecting when it was too close to the startingRoom's window when doing the intro

Also fixed the menu not properly disabeling the video/sens panels before they are animated off the screen.