Updated how the player handles travelling from inside to outside
Once the startingRoom's glass shatters, a value in the playerController will get ticked. This value indicates that a function will get called every frame to check the player and their camera's position, which determines when the player and their camera are outside, sepperatly. Once the player is far enough away from the window, the function stops constantly updating itself.

Note: A problem found is that the player's body can block the player's camera if they move at a certain angle. This is due to the player's camera