Adjusted some bugs and efficiency with the playerController
Any calls to get a layer using NameToLayer has been changed to get the direct integer.

The player now tracks their current state as an actual PlayerState value and no longer an integer of the PlayerState. This requires a new state of NULL so that when calling ChangeState to set the state, we have a default state to transition from, no matter the state we are changing into.

Setting the player's layer to ignoreRaycast will allow the sun's flare to ignore the collider of the player's model. This will prevent the sun's rays from diminishing if the player's camera leaves the player's model.