Implemented a second font to the menu. Adjusted how hovering over buttons work.
A second font has been added to the menu that will be used only with the main menu buttons. This new font has increased the width ratios of the buttons.

When setting up a button, it will now allign to the left. The hover value of each button will now reduce the width of the button if it is not hovered/selected. That way, when hovering/selecting a button, the width will increase, which will increase the button's font size as the text tries to fill the space.