Added a value that unified the sizes of the buttons
buttonSizeMod will control the overall size of the buttons in the menu. The start and quit buttons now both have a unique modifier on their size, making the start button larger and the quit button smaller. When handling the size of the quit button, do not include the unique size mod to it. This is so that the button is an equal distance away from the other buttons, but still has a smaller size.