Added the basis for a mosue click and implemented out of menu mouse lock
When the user clicks on a button in the menu, a request to play a mouse click sound is sent.

Handling the mouse lock state has been adjusted. Now the mouse will go into the locked state when the player has control, which will hide the mouse. When returning to the menu, it will preserve what the previous mouse state was, ie confined or open.