Added a sun material and updated slider visuals
The sun now has it's own material to make it blend in better with the sunflare and look more like a sun when the sun flare starts up.

The slider has it's visuals updated. It's now more clear as to what side will increase the sensitivity and the handle of the slider is that of a mouse to fully connect it's usage.