Changing rooms will only disable/enable the portals
When changing rooms, the portals of the puzzle rooms and the portals of the windows of the waitingRooms will be the objects that are enabled/diabled. For now, the entire startingRoom will be enabled/disabled.

This action however still causes lag, so next update will aim to reduce said lag even further.