Loading and unlopading new rooms is now lagless
When loading and unloading rooms, the portals in said rooms will run a function that will disable or enable the portal's meshRenderer to prevent the player from rendering any portals that are outside of their view. This method makes travelling between rooms lagless instead of the previous object enabling/disabling.