Fixed an issue with the stairs creation and a few visuals in puzzleRoom 7
The stairs objects were not deleting their previously created points of interest, so there was over 10 thousand useless transforms in the game. The stairs now delete their PoI before they create any more.

The player was able to get trapped in the puzzleRoom 7 by falling onto the forward wall bellow the tube/hall path. We added a set of stairs to let them walk out now.

The player's custom RayTrace function has been changed slightly, but the issue of having the player pass through a portal can still occur.

Note: There is still a sudden burst of lag that occurs when moving between the final 3 rooms. Testing points to the portals and how they react to the player camera.