Each portal now has a unique depth value to improve performance
The issue when moving around in a portal heavy room was determined to be an unallocated value taking up CPU cycles (this was found using the profiler). To prevent this from occuring, each portalSet now has it's own depth value which will prevent them from rendering any further if it's impossible for them to be seen. This has improved the performance, and so far I was unable to see any non-rendered portals in puzzleRoom 7 (But further testing should be done).