Fixed the player movement method to solve the issue of not teleporting
The issue that was causing the player to fail to teleport is due to how the player was being moved by the physics system and not using the custom raytrace. This however kept the player from doing wierd movements and fitting between wierd places. Currently there are no known corners the player can phase through due to the pillars thankfully.

Note: Currently, when trying to enter a portal that is too close to a wall/object, the player will collide with the wall through the portal. Travelling while outside and moving into slanted terrain has a wierd effect that needs to be looked at.