Major improvements to the way the portal renders it's view makes it nearly perfect
Updated the rendertextures used by the portals. By increasing their render depth and having a dynamic resolution, getting close to a portal will now not cause jittering and rendering issues from the camera's side.

Currently, the biggets visual difference when teleporting is the edges of the portals wn teleporting. This issue can be solved by implementing a new script, but the issue is very minor and may not be neccesairy.