Updated how the player falls into the void of an open puzzleRoom
When the player falls past the play area of the puzzleRoom. they will now gain speed relative to how far they are from the play area. This now makes it so the player spends less time falling with nothing to look at, while the quick increase and decrease of the player's speed is masked by the visual effects and the lack of landmarks.

Also, the toggle button's checkmarks in the video panel now adjust to reflect the screen resolution as it changes after initial launch.