GlobalTerrainController now requests each room to change their textures
The GlobalTerrainController is now called by an attachedRoom to change the textures of a series of rooms that are position behind the attachedRoom that called the function.

Currently, the startingRoom can handle this request by having it's materials created instead of referenced (now it requires references to textures). This will allow the room to change the textures of it's material without effecting other rooms.

Also, prevented the player form firing a movementVector raytrace while the movementVector has a magnitude of zero.