Recommitted the previous commit. WaitingRoom now has it's own materials.
I accidentally deleted the testArea scene on the previous commit, and it was too large to re-add alone. This commit is the same as the previous commit that was undone, but has the added functionnality of unique materials for WaitingRooms and AttachedRooms.

Currently, the startingRoom and the waitingRooms and their attachedRooms all have their own materials to allow us to have unique textures for each waitingRoom.