Each puzzleRoom and it's objects now use it's own materials
Each puzzleRoom now has it's own materials, which are now used by each object in the given room.

This lets the puzzleRoom adjust the textures of the cubes and pillars that are only in it's room. Currently, we only set the textures to null (but it does work for each object and each room individually).