Started implementing the changing textures for the puzzleRoom
The puzzleRoom now tracks a series of textures which will be the combination of the two given texutres. Currently, when the textures of the room change, it only changes the first texture of the walls. The altTexturePriority will decide how much of either texture is used in the combination process.

The action of combining the textures is expensive, so this may need to be done on startup instead of on change.

Also, to be able to access the pixels of a texture, it must be open for reading/writting, which means it will not be compressed. This may cause a large increase in the memory required to run the game, so next build may have a large increase in memory needed to run. In the profiler, this version requires 100mb for textures on startup.