PuzzleRoom texture changing now uses precomputed textures
The previous intended method of combining two textures was an expensive calculation for both CPU and memory. The backup method uses predetermined textures that will replace the current texture used on the given materials.

Currently, each puzzleRoom has a unique alternative texture for each of it's used materials that it can switch to on command.