Adjusted the outside terrain and window placement
The outside terrain will now be disabled once the player loses sight/waiting room unloads the startingRoom. This is because the player can still interract with the terrain as it is popping into the puzzleRooms.

Also, the camera is now placed in the center of a terrainChunk to prevent the player from moving around and unloading a chunk. Preferably, we would want the chunks to not change while the player is not yet outside, but this works for now.