Adjusted the outside texture and sounds
The outside skySphere now has a better blend into the terrain. A new audio clip is introduced and is used as the music during the main menu on startup (outside fields background sounds). Once the player presses the start button, the song will switch to the original starting music, Gymnopedie.