Fixed a bug that caused the music to incorrectly and fail to loop
When there was no music playing, the playerSounds script would try to play music on a delay. However, since the music would not actually start on the call and this check occured on every update call, the music would never play.

Also, due to how it checked whether the music was playing, if the muted version of the song ended, then the upgraded version would be forced to loop also. This caused the starting song to loop early due to the muted song being the outside background noise.

Currently, any song can be playing on either the muted or upgraded versions and they will both loop independently. This includes when the player has entered the outside (when the song loops, it will also change songs succesfully).

Also, removed Another Sky as a song. It doesn't sound great while muted.